usb: misc: usb3503: Force late initialization
While not working as an I2C device, USB3503 chip needs to be reset
after the USB PHY has been initialized for this to work properly.

Currently there is no other way to ensure that USB3503 chip is probed
after the USB PHY has been initialized, hence the last resort.

Signed-off-by: Tushar Behera <tushar.behera@linaro.org>