clockevents: add ->set_dev_mode() to 'struct clock_event_device'
When no timers/hrtimers are pending, the expiry time is set to a special value:
'KTIME_MAX'.

This would normally happen on idle and full dynticks target in both LOW/HIGH
resolution modes.

When expiry is set to KTIME_MAX, in highres mode we cancel the 'tick-sched'
hrtimer and in lowres mode we skip reprogramming clockevent device. But, the
clockevent device is already reprogrammed from tick-handler.

So, the clockevent device will fire atleast one more time.

In highres mode we consider it as a spurious interrupt and simply return from
hrtimer_interrupt().

In lowres mode, we haven't cancelled the event (like we cancelled hrtimer in
highres) and so tick_nohz_handler() will keep on rescheduling tick. One way out
to fix this issue is not to reschedule tick if 'tick_stopped' is set. This will
be addressed by a separate patchset.

As the clock event device is programmed in ONESHOT mode it should fire one more
time (unnecessarily). But, many implementations (like arm_arch_timer, powerpc,
etc.) only support PERIODIC mode and their drivers emulate ONESHOT over that.
And so we will keep getting these spurious interrupts at tick-rate.

In order to avoid these spurious interrupts/wakeups, the clockevent device
should be stopped, or its interrupts should be masked.

A simple (yet hacky) solution to get this fixed could be: update
hrtimer_force_reprogram() to always reprogram clockevent device and update
clockevent drivers to STOP generating events (or delay it to max time) when
'expires' is set to KTIME_MAX. But the drawback here is that every clockevent
driver has to be hacked for this particular case and its very easy for new ones
to miss this.

However, Thomas suggested to add an optional mode ONESHOT_STOPPED to solve this
problem: lkml.org/lkml/2014/5/9/508

With this proposal, introducing a new ONESHOT_STOPPED mode would require the
core to know whether the platform implements this mode so it could be
reprogrammed later.

In order for the core to tell if the mode is implemented, ->set_mode() callback
needs to be able to return success or failure.

To change return type of ->set_mode(), Thomas suggested to implement another
callback: ->set_dev_mode(), with return type 'int'. We can then convert
clockevent drivers to use this interface instead of existing ->set_mode() and
then finally remove ->set_mode()'s support from core.

This patch adds support for ->set_dev_mode() in clockevents core.

For now, just add a WARN_ON_ONCE() to callers of ->set_dev_mode() as none of
currently implemented modes are expected/allowed to fail.

Reviewed-by: Preeti U Murthy <preeti@linux.vnet.ibm.com>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
[
Backported from viresh branch tick/oneshot-stopped.
Noticed below hunk whose place holder missing in 3.14
so ignoring them.

file kernel/time/clockevent.c

@@ -446,8 +454,14 @@ int __clockevents_update_freq(struct clock_event_device *dev, u32 freq)
if (dev->mode == CLOCK_EVT_MODE_ONESHOT)
	return clockevents_program_event(dev, dev->next_event, false);

-   if (dev->mode == CLOCK_EVT_MODE_PERIODIC)
-           dev->set_mode(CLOCK_EVT_MODE_PERIODIC, dev);
+   if (dev->mode == CLOCK_EVT_MODE_PERIODIC) {
+           /* Shouldn't fail while switching to PERIODIC MODE */
+           if (dev->set_dev_mode)
+                   WARN_ON_ONCE(dev->set_dev_mode(CLOCK_EVT_MODE_PERIODIC,
+                                                  dev));
+           else
+                   dev->set_mode(CLOCK_EVT_MODE_PERIODIC, dev);
+   }

]

Signed-off-by: Santosh Shukla <santosh.shukla@linaro.org>