clockevents: Stop unused clockevent device
Clockevent device can now be stopped by switching to ONESHOT_STOPPED mode, to
avoid getting spurious interrupts on a tickless CPU.

This patch switches mode to ONESHOT_STOPPED at three different places and
following is the reasoning behind them.

1.) NOHZ_MODE_LOWRES

Timers & hrtimers are dependent on tick for their working in this mode and the
only place from where clockevent device is programmed is the tick-code.

So, we only need to switch clockevent device to ONESHOT_STOPPED mode once ticks
aren't required anymore. And the only call site is: tick_nohz_stop_sched_tick().

In LOWRES mode we skip reprogramming the clockevent device here if expires ==
KTIME_MAX. In addition to that we must also switch the clockevent device to
ONESHOT_STOPPED mode to avoid all spurious interrupts that may follow.

2.) NOHZ_MODE_HIGHRES

Tick & timers are dependent on hrtimers for their working in this mode and the
only place from where clockevent device is programmed is the hrtimer-code.

There are two places here from which we reprogram the clockevent device or skip
reprogramming it on expires == KTIME_MAX.

Instead of skipping reprogramming the clockevent device, also switch its mode to
ONESHOT_STOPPED so that it doesn't generate any spurious interrupts.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
[forward port to 3.18
 - manually applied few patch.
]
Signed-off-by: Santosh Shukla <santosh.shukla@linaro.org>